requirements.txt
playwright==1.58.0
pytest==9.0.2
pytest-html==4.2.0
pytest-xdist==3.8.0
openai==2.30.0
requests==2.33.1
allure-pytest==2.15.3
appium-python-client==5.3.0
python-dotenv==1.0.1
curl_cffi==0.15.0
